requirements.txt
flask
flask-cors
supabase
pdfplumber
pymupdf
pytesseract
pillow
gunicorn
